moved some methods from env to agent
