+++ rl_equation_solver/config.py
@@ -5,7 +5,7 @@ class Config:
     """Model configuration"""
 
     # BATCH_SIZE is the number of Experience sampled from the replay buffer
-    BATCH_SIZE = 256
+    BATCH_SIZE = 64
     # GAMMA is the discount factor
     GAMMA = 0.99
     # EPSILON_START is the starting value of epsilon
@@ -14,7 +14,7 @@ class Config:
     EPSILON_END = 0.05
     # EPSILON_DECAY controls the rate of exponential decay of epsilon, higher
     # means a slower decay
-    EPSILON_DECAY = 10
+    EPSILON_DECAY = 20
     # TAU is the update rate of the target network
     TAU = 0.005
     # LR is the learning rate of the AdamW optimizer
